+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## NEXT
+## 1.0
 
 - Adds `scripts/run.sh` to run this action with a bash script.
 - **BREAKING**: `fail-on` now defaults to `'warning'`.